New criteria to find config folder
Now the directory hierarchy is searched up for a ".bundle" dir and if
not foun, current folder is used. This is more intuitive than choosing
the configuration location than using the folder of the Gemfile being
used, but also it will allow us to respect the Gemfile configured via
settings since the new algorithm does not need a Gemfile.